Finished boss skydive attack, for some reason when I change the tag of one prefab, it changes all the other prefabs to the same tag, so the boss still ends up colliding with the EOD that it makes upon landing

+++ Assets/Scripts/Game/EntityBase.cs
@@ -11,8 +11,8 @@ namespace BossArena.game
     public abstract class EntityBase : NetworkBehaviour
     {
         [SerializeField]
-        public int MaxHealth { get; protected set; }
-        public int CurrentHealth { get; set; }
+        public float MaxHealth { get; protected set; }
+        public float CurrentHealth { get; set; }
         public bool IsAlive { get; protected set; }
 
         public EntityState State { get; set; }
@@ -37,7 +37,7 @@ namespace BossArena.game
 
         protected abstract void LateUpdate();
 
-        protected void SetHealth(int health)
+        protected void SetHealth(float health)
         {
             MaxHealth = health;
             CurrentHealth = health;
